requirements.txt
streamlit==1.38.0
python-dotenv>=1.0.1
pydantic>=2.7.0
rank-bm25>=0.2.2
rapidfuzz>=3.9.0
pypdf>=4.2.0
openai>=1.51.0
loguru>=0.7.2
numpy>=1.26.0
